+++ include/gp_nvvideo_decoder.h
@@ -23,59 +23,90 @@
 #include "context.h"
 #include "gp_beader.h"
 #include "gp_circular_buffer.h"
+#include "gp_threadpool.h"
 #include "gplayer.h"
 
-#define MICROSECOND_UNIT 1000000
-#define CHUNK_SIZE 4000000
-#define MIN(a, b) (((a) < (b)) ? (a) : (b))
-
-#define IS_NAL_UNIT_START(buffer_ptr) \
-    (!buffer_ptr[0] && !buffer_ptr[1] && !buffer_ptr[2] && (buffer_ptr[3] == 1))
-
-#define IS_NAL_UNIT_START1(buffer_ptr) \
-    (!buffer_ptr[0] && !buffer_ptr[1] && (buffer_ptr[2] == 1))
-
-#define H264_NAL_UNIT_CODED_SLICE 1
-#define H264_NAL_UNIT_CODED_SLICE_IDR 5
-
-#define HEVC_NUT_TRAIL_N 0
-#define HEVC_NUT_RASL_R 9
-#define HEVC_NUT_BLA_W_LP 16
-#define HEVC_NUT_CRA_NUT 21
-
-#define IVF_FILE_HDR_SIZE 32
-#define IVF_FRAME_HDR_SIZE 12
-
-#define IS_H264_NAL_CODED_SLICE(buffer_ptr) \
-    ((buffer_ptr[0] & 0x1F) == H264_NAL_UNIT_CODED_SLICE)
-#define IS_H264_NAL_CODED_SLICE_IDR(buffer_ptr) \
-    ((buffer_ptr[0] & 0x1F) == H264_NAL_UNIT_CODED_SLICE_IDR)
-
-#define GET_H265_NAL_UNIT_TYPE(buffer_ptr) ((buffer_ptr[0] & 0x7E) >> 1)
+#define USE_NVBUF_TRANSFORM_API
+#define MAX_BUFFERS 32
 
 namespace GPlayer {
 
-using namespace std;
+typedef struct : public context_t {
+    NvVideoDecoder* dec;
+    NvVideoConverter* conv;
+    uint32_t decoder_pixfmt;
+    bool fullscreen;
+    uint32_t window_height;
+    uint32_t window_width;
+    uint32_t window_x;
+    uint32_t window_y;
+    uint32_t out_pixfmt;
+    uint32_t video_height;
+    uint32_t video_width;
+    uint32_t display_height;
+    uint32_t display_width;
+    float fps;
+
+    bool disable_dpb;
+
+    bool input_nalu;
+
+    bool copy_timestamp;
+    bool flag_copyts;
+    uint32_t start_ts;
+    float dec_fps;
+    uint64_t timestamp;
+    uint64_t timestampincr;
+
+    bool stats;
+
+    bool enable_metadata;
+    // bool bLoop;
+    // bool bQueue;
+    bool enable_input_metadata;
+    enum v4l2_skip_frames_type skip_frames;
+    enum v4l2_memory output_plane_mem_type;
+    enum v4l2_memory capture_plane_mem_type;
+#ifndef USE_NVBUF_TRANSFORM_API
+    enum v4l2_yuv_rescale_method rescale_method;
+#endif
+
+    std::queue<NvBuffer*>* conv_output_plane_buf_queue;
+    pthread_mutex_t queue_lock;
+    pthread_cond_t queue_cond;
+
+    pthread_t dec_capture_loop;  // Decoder capture thread, created if running
+                                 // in blocking mode.
+    bool got_error;
+    bool got_eos;
+    bool vp9_file_header_flag;
+    bool vp8_file_header_flag;
+    int dst_dma_fd;
+    int dmabuff_fd[MAX_BUFFERS];
+    int numCapBuffers;
+    int loop_count;
+    int max_perf;
+    int extra_cap_plane_buffer;
+} VideoDecodeContext_T;
 
 class GPNvVideoDecoder : public IBeader {
 private:
     GPNvVideoDecoder();
 
 public:
-    explicit GPNvVideoDecoder(const shared_ptr<VideoDecodeContext_T> context);
+    explicit GPNvVideoDecoder(
+        const std::shared_ptr<VideoDecodeContext_T> context);
     ~GPNvVideoDecoder();
 
     std::string GetInfo() const override;
-
     void Process(GPData* data);
+    int Proc() override;
+    bool HasProc() override { return true; };
 
 private:
     int read_decoder_input_nalu(NvBuffer* buffer);
-
     int read_decoder_input_chunk(NvBuffer* buffer);
-
     int read_vpx_decoder_input_chunk(NvBuffer* buffer);
-
     void Abort();
 
 #ifndef USE_NVBUF_TRANSFORM_API
@@ -92,7 +123,6 @@ private:
 
     int report_input_metadata(
         v4l2_ctrl_videodec_inputbuf_metadata* input_metadata);
-
     void report_metadata(v4l2_ctrl_videodec_outputbuf_metadata* metadata);
 
 #ifndef USE_NVBUF_TRANSFORM_API
@@ -100,27 +130,21 @@ private:
 #endif
 
     void query_and_set_capture();
-
-    static void* decoder_pollthread_fcn(void* arg);
-
-    static void* dec_capture_loop_fcn(void* arg);
-
+    void* decoder_pollthread_fcn(void);
     void set_defaults();
-
-    bool decoder_proc_nonblocking(bool eos,
-                                  uint32_t current_file,
-                                  int current_loop);
-    bool decoder_proc_blocking(bool eos,
-                               uint32_t current_file,
-                               int current_loop);
-    int Proc() override;
-    bool HasProc() override { return true; };
+    bool decoder_proc_nonblocking(bool eos);
+    void ProcessData();
 
 private:
     std::shared_ptr<VideoDecodeContext_T> ctx_;
+    std::thread decoder_poll_thread_;
     gp_circular_buffer<uint8_t> buffer_;
+    bool decoder_is_availiable_;
     std::mutex buffer_mutex_;
-    std::condition_variable thread_condition_;
+    std::mutex decoder_poll_mutex_;
+    std::mutex process_mutex_;
+    std::condition_variable decoder_poll_condition_;
+    std::condition_variable process_condition_;
 };
 
 };  // namespace GPlayer
